organized and optimized code
