+++ src/parser.ts
@@ -1,7 +1,14 @@
 import { Command, Option } from "commander";
 
-import Action from "./actions.js";
 import config from "./config.js";
+
+import Balance from "./Commands/Balance.js";
+import Block from "./Commands/Block.js";
+import Blocknumber from "./Commands/Blocknumber.js";
+import Transaction from "./Commands/Transaction.js";
+import Compile from "./Commands/Compile.js";
+import Deploy from "./Commands/Deploy.js";
+import Interact from "./Commands/Interact.js";
 
 const cli = new Command("eth").version(config.version);
 
@@ -16,31 +23,27 @@ const parse = () => {
         .description("get balance of address")
         .requiredOption("--address <string>", "account or contract address")
         .action((args) => {
-            const network = cli.opts().network;
-            new Action(network).showBalance(args.address);
+            new Balance(cli.opts().network).showBalance(args.address);
         });
 
     cli.command("blocknumber")
         .description("get latest block number")
         .action(() => {
-            const network = cli.opts().network;
-            new Action(network).showBlockNumber();
+            new Blocknumber(cli.opts().network).showBlockNumber();
         });
 
     cli.command("block")
         .description("get block data")
         .requiredOption("--number <block number>", "get data of block number")
         .action((args) => {
-            const network = cli.opts().network;
-            new Action(network).showBlock(parseInt(args.number));
+            new Block(cli.opts().network).showBlock(parseInt(args.number));
         });
 
     cli.command("transaction")
         .description("get transaction data")
         .requiredOption("--hash <transaction hash>", "transaction hash")
         .action((args) => {
-            const network = cli.opts().network;
-            new Action(network).showTransaction(args.hash);
+            new Transaction(cli.opts().network).showTransaction(args.hash);
         });
 
     cli.command("compile")
@@ -52,8 +55,7 @@ const parse = () => {
             "path to solidity smart contract source code"
         )
         .action((args) => {
-            const network = cli.opts().network;
-            new Action(network).compile(args.src);
+            new Compile(cli.opts().network).compile(args.src);
         });
 
     cli.command("deploy")
@@ -65,8 +67,11 @@ const parse = () => {
         .requiredOption("--abi <abi path>", "path to contract abi")
         .requiredOption("--key <private key>", "private key")
         .action((args) => {
-            const network = cli.opts().network;
-            new Action(network).deploy(args.bytecode, args.abi, args.key);
+            new Deploy(cli.opts().network).deploy(
+                args.bytecode,
+                args.abi,
+                args.key
+            );
         });
 
     cli.command("interact")
@@ -82,8 +87,7 @@ const parse = () => {
             "private key is needed to call state changing methods"
         )
         .action((args) => {
-            const network = cli.opts().network;
-            new Action(network).interact(
+            new Interact(cli.opts().network).interact(
                 args.contract,
                 args.abi,
                 args.method,
